more refactoring and documentation

+++ server/api/post/postFunctions.js
@@ -1,8 +1,17 @@
-const User = require('../auth/userModel');
-
-exports.deletePostandComments = async function deletePostandComments(postId, Post, commentController) {
+/**
+ * This function finds a post by its postId and deletes the post as well as calling the deleteComments function.
+ * @param {string} postId - Mongo Id of the post to be deleted
+ * @param {Object} Post - Either the Post model or a mock for testing
+ * @param {Object} commentController - The commentController with functions for the comments
+ */
+exports.deletePostandComments = async function deletePostandComments(postId, userId, Post, commentFunctions, User) {
   try {
-    const deleteComments = await commentController.deleteComments(postId);
+    const post = await Post.findOne({ postId });
+    const user = await User.findOne({ username: post.username });
+    if (user._id != userId) {
+      return { success: false, msg: 'Unauthorized' };
+    }
+    const deleteComments = await commentFunctions.deleteComments(postId);
     const deletePost = await Post.findOneAndRemove({ _id: postId });
     if (deleteComments && deletePost) {
       return true;
@@ -14,6 +23,11 @@ exports.deletePostandComments = async function deletePostandComments(postId, Pos
   }
 };
 
+/**
+ * This function will return the last 10 post, if a lastPostDate is passed in it will get the last 10 post before that date.
+ * @param {Date} lastPostDate - Mainly to be used to get more posts for the feed, pass in the date of the last image on the current feed.
+ * @param {Object} Post - Either the post model or a mock for testing
+ */
 exports.latestPost = async function latestPost(lastPostDate, Post) {
   try {
     let posts = [];
@@ -29,7 +43,13 @@ exports.latestPost = async function latestPost(lastPostDate, Post) {
   }
 };
 
-exports.newPost = async function newComment(imgUrl, userId, Post) {
+/**
+ * Creates a new post
+ * @param {Id} imgUrl - the Url(might change later to database id) of the new image
+ * @param {Id} userId - the Id of the user, the auth middelware inserts this into req.user.id
+ * @param {Object} Post - Either the post model or a mock for testing
+ */
+exports.newPost = async function newComment(imgUrl, userId, Post, User) {
   try {
     const user = await User.findOne({ _id: userId });
     if (user) {
